backend/requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
Pygments==2.17.2
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
pytz==2024.1
rich==13.7.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.9.0
Werkzeug==3.0.1
